Reorganize some big files
